requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.0
langchain-core>=0.1.0
requests>=2.25.0
